LayoutTests/fast/multicol/hit-test-above-or-below-expected.txt

123
abc
def
ghi
jkl
mno
Character at 150, 25 is 1 as expected.
Character at 350, 25 is d as expected.
Character at 550, 25 is j as expected.
Character at 750, 25 is m as expected.
Character at 150, 275 is d as expected.
Character at 350, 275 is j as expected.
Character at 550, 275 is m as expected.
Character at 750, 275 is null as expected.
Character at 150, 475 is d as expected.
Character at 350, 475 is j as expected.
Character at 550, 475 is m as expected.
Character at 750, 475 is null as expected.
